compose splitted for a better integration in other projects
The main compose has been splitted in multiple compose, one for each service. This repo can now integrates with another projects. Also the warnings are removed.
A network is created when running the up.sh script so that all the service can comunicate with each other. The network can be deleted when executing the down.sh script with the -n or --networks option.